complete Zipcode preferences UI
