+++ client/src/components/pages/ProfilePage/Profile/PreferenceTab/PreferenceTab.js
@@ -1,37 +1,13 @@
 import React, { useState } from "react"
-import * as Fields from "./PreferenceFields"
-import FormWrapper from "components/ui/AccountFormWrapper/AccountFormWrapper"
+import Gender from "./Gender"
+import ZipCodes from "./ZipCodes"
 
-const PreferenceTab = ({
-  updateAccountDetails,
-  accountDetails,
-  currentUser,
-  formEntries,
-  setFormEntries,
-  handleSubmit,
-  handleChange,
-  madeChange,
-  setMadeChange,
-}) => {
-  return (
-    <>
-      <Fields.GenderButton
-        madeAChange={() => setMadeChange(true)}
-        formEntries={formEntries}
-        setFormEntries={setFormEntries}
-      />
-
-      <Fields.ZipCodeList
-        formEntries={formEntries}
-        setFormEntries={setFormEntries}
-        handleSubmit={handleSubmit}
-      />
-    </>
-  )
-}
-
-const WrappedTab = (props) => (
-  <FormWrapper FieldsComponent={PreferenceTab} {...props} />
+const PreferenceTab = () => (
+  <>
+    <Gender />
+    <br />
+    <ZipCodes />
+  </>
 )
 
-export default WrappedTab
+export default PreferenceTab
